+++ modules/contracts/src.ts/services/ethReader.ts
@@ -30,6 +30,7 @@ import {
 import axios from "axios";
 import { encodeBalance, encodeTransferResolver, encodeTransferState } from "@connext/vector-utils";
 import { BigNumber } from "@ethersproject/bignumber";
+import { parseUnits } from "@ethersproject/units";
 import { AddressZero, HashZero } from "@ethersproject/constants";
 import { Contract } from "@ethersproject/contracts";
 import { JsonRpcProvider, TransactionRequest } from "@ethersproject/providers";
@@ -37,6 +38,9 @@ import pino from "pino";
 
 import { ChannelFactory, ChannelMastercopy, TransferDefinition, TransferRegistry, VectorChannel } from "../artifacts";
 import { Evt } from "evt";
+
+export const MIN_GAS_PRICE = parseUnits("5", "gwei");
+export const BUMP_GAS_PRICE = 30; // 30% bump gas price
 
 // https://github.com/rustwasm/wasm-bindgen/issues/700#issuecomment-419708471
 const execEvmBytecode = (bytecode: string, payload: string): Uint8Array =>
@@ -84,6 +88,11 @@ export class EthereumChainReader implements IVectorChainReader {
       ChainError
     >
   > {
+    const provider = this.chainProviders[chainId];
+    if (!provider) {
+      return Result.fail(new ChainError(ChainError.reasons.ProviderNotFound));
+    }
+
     return await this.retryWrapper<
       | boolean
       | {
@@ -91,7 +100,7 @@ export class EthereumChainReader implements IVectorChainReader {
           currentBlock: string;
           highestBlock: string;
         }
-    >(chainId, async (provider: JsonRpcProvider) => {
+    >(chainId, async () => {
       try {
         const res = await provider.send("eth_syncing", []);
         return Result.ok(res);
@@ -105,16 +114,22 @@ export class EthereumChainReader implements IVectorChainReader {
     channelAddress: string,
     chainId: number,
   ): Promise<Result<ChannelDispute | undefined, ChainError>> {
-    return await this.retryWrapper<ChannelDispute | undefined>(chainId, async (provider: JsonRpcProvider) => {
+    const provider = this.chainProviders[chainId];
+    if (!provider) {
+      return Result.fail(new ChainError(ChainError.reasons.ProviderNotFound));
+    }
+
+    const code = await this.getCode(channelAddress, chainId);
+    if (code.isError) {
+      return Result.fail(code.getError()!);
+    }
+
+    if (code.getValue() === "0x" || code.getValue() === undefined) {
+      // channel is not deployed
+      return Result.ok(undefined);
+    }
+    return await this.retryWrapper<ChannelDispute | undefined>(chainId, async () => {
       try {
-        const code = await this.getCode(channelAddress, chainId);
-        if (code.isError) {
-          return Result.fail(code.getError()!);
-        }
-        if (code.getValue() === "0x") {
-          // channel is not deployed
-          return Result.ok(undefined);
-        }
         const dispute = await new Contract(channelAddress, VectorChannel.abi, provider).getChannelDispute();
         if (dispute.channelStateHash === HashZero) {
           return Result.ok(undefined);
@@ -138,7 +153,7 @@ export class EthereumChainReader implements IVectorChainReader {
     chainId: number,
     bytecode?: string,
   ): Promise<Result<RegisteredTransfer, ChainError>> {
-    return await this.retryWrapper<RegisteredTransfer>(chainId, async (provider: JsonRpcProvider) => {
+    return await this.retryWrapper<RegisteredTransfer>(chainId, async () => {
       let registry = this.transferRegistries.get(chainId.toString())!;
       if (!this.transferRegistries.has(chainId.toString())) {
         // Registry for chain not loaded, load into memory
@@ -169,7 +184,7 @@ export class EthereumChainReader implements IVectorChainReader {
     chainId: number,
     bytecode?: string,
   ): Promise<Result<RegisteredTransfer, ChainError>> {
-    return await this.retryWrapper<RegisteredTransfer>(chainId, async (provider: JsonRpcProvider) => {
+    return await this.retryWrapper<RegisteredTransfer>(chainId, async () => {
       let registry = this.transferRegistries.get(chainId.toString());
       if (!registry) {
         // Registry for chain not loaded, load into memory
@@ -214,7 +229,11 @@ export class EthereumChainReader implements IVectorChainReader {
   }
 
   async getChannelFactoryBytecode(channelFactoryAddress: string, chainId: number): Promise<Result<string, ChainError>> {
-    return await this.retryWrapper<string>(chainId, async (provider: JsonRpcProvider) => {
+    const provider = this.chainProviders[chainId];
+    if (!provider) {
+      return Result.fail(new ChainError(ChainError.reasons.ProviderNotFound));
+    }
+    return await this.retryWrapper<string>(chainId, async () => {
       try {
         const factory = new Contract(channelFactoryAddress, ChannelFactory.abi, provider);
         const proxyBytecode = await factory.getProxyCreationCode();
@@ -229,7 +248,11 @@ export class EthereumChainReader implements IVectorChainReader {
     channelFactoryAddress: string,
     chainId: number,
   ): Promise<Result<string, ChainError>> {
-    return await this.retryWrapper<string>(chainId, async (provider: JsonRpcProvider) => {
+    const provider = this.chainProviders[chainId];
+    if (!provider) {
+      return Result.fail(new ChainError(ChainError.reasons.ProviderNotFound));
+    }
+    return await this.retryWrapper<string>(chainId, async () => {
       try {
         const factory = new Contract(channelFactoryAddress, ChannelFactory.abi, provider);
         const mastercopy = await factory.getMastercopy();
@@ -245,7 +268,11 @@ export class EthereumChainReader implements IVectorChainReader {
     chainId: number,
     assetId: string,
   ): Promise<Result<BigNumber, ChainError>> {
-    return await this.retryWrapper<BigNumber>(chainId, async (provider: JsonRpcProvider) => {
+    const provider = this.chainProviders[chainId];
+    if (!provider) {
+      return Result.fail(new ChainError(ChainError.reasons.ProviderNotFound));
+    }
+    return await this.retryWrapper<BigNumber>(chainId, async () => {
       const code = await this.getCode(channelAddress, chainId);
       if (code.isError) {
         return Result.fail(code.getError()!);
@@ -270,7 +297,11 @@ export class EthereumChainReader implements IVectorChainReader {
     chainId: number,
     assetId: string,
   ): Promise<Result<BigNumber, ChainError>> {
-    return await this.retryWrapper<BigNumber>(chainId, async (provider: JsonRpcProvider) => {
+    const provider = this.chainProviders[chainId];
+    if (!provider) {
+      return Result.fail(new ChainError(ChainError.reasons.ProviderNotFound));
+    }
+    return await this.retryWrapper<BigNumber>(chainId, async () => {
       const code = await this.getCode(channelAddress, chainId);
       if (code.isError) {
         return Result.fail(code.getError()!);
@@ -298,7 +329,11 @@ export class EthereumChainReader implements IVectorChainReader {
     chainId: number,
     bytecode?: string,
   ): Promise<Result<boolean, ChainError>> {
-    return await this.retryWrapper<boolean>(chainId, async (provider: JsonRpcProvider) => {
+    const provider = this.chainProviders[chainId];
+    if (!provider) {
+      return Result.fail(new ChainError(ChainError.reasons.ProviderNotFound));
+    }
+    return await this.retryWrapper<boolean>(chainId, async () => {
       // Get encoding
       const registryRes = await this.getRegisteredTransferByDefinition(
         transferDefinition,
@@ -345,7 +380,11 @@ export class EthereumChainReader implements IVectorChainReader {
   }
 
   async resolve(transfer: FullTransferState, chainId: number, bytecode?: string): Promise<Result<Balance, ChainError>> {
-    return await this.retryWrapper<Balance>(chainId, async (provider: JsonRpcProvider) => {
+    const provider = this.chainProviders[chainId];
+    if (!provider) {
+      return Result.fail(new ChainError(ChainError.reasons.ProviderNotFound));
+    }
+    return await this.retryWrapper<Balance>(chainId, async () => {
       // Try to encode
       let encodedState: string;
       let encodedResolver: string;
@@ -395,7 +434,11 @@ export class EthereumChainReader implements IVectorChainReader {
     channelFactoryAddress: string,
     chainId: number,
   ): Promise<Result<string, ChainError>> {
-    return await this.retryWrapper<string>(chainId, async (provider: JsonRpcProvider) => {
+    const provider = this.chainProviders[chainId];
+    if (!provider) {
+      return Result.fail(new ChainError(ChainError.reasons.ProviderNotFound));
+    }
+    return await this.retryWrapper<string>(chainId, async () => {
       const channelFactory = new Contract(channelFactoryAddress, ChannelFactory.abi, provider);
       try {
         const derivedAddress = await channelFactory.getChannelAddress(alice, bob);
@@ -407,7 +450,11 @@ export class EthereumChainReader implements IVectorChainReader {
   }
 
   async getCode(address: string, chainId: number): Promise<Result<string, ChainError>> {
-    return await this.retryWrapper<string>(chainId, async (provider: JsonRpcProvider) => {
+    const provider = this.chainProviders[chainId];
+    if (!provider) {
+      return Result.fail(new ChainError(ChainError.reasons.ProviderNotFound));
+    }
+    return await this.retryWrapper<string>(chainId, async () => {
       try {
         const code = await provider.getCode(address);
         return Result.ok(code);
@@ -418,7 +465,11 @@ export class EthereumChainReader implements IVectorChainReader {
   }
 
   async getBlockNumber(chainId: number): Promise<Result<number, ChainError>> {
-    return await this.retryWrapper<number>(chainId, async (provider: JsonRpcProvider) => {
+    const provider = this.chainProviders[chainId];
+    if (!provider) {
+      return Result.fail(new ChainError(ChainError.reasons.ProviderNotFound));
+    }
+    return await this.retryWrapper<number>(chainId, async () => {
       try {
         const blockNumber = await provider.getBlockNumber();
         return Result.ok(blockNumber);
@@ -429,7 +480,11 @@ export class EthereumChainReader implements IVectorChainReader {
   }
 
   async getGasPrice(chainId: number): Promise<Result<BigNumber, ChainError>> {
-    return await this.retryWrapper<BigNumber>(chainId, async (provider: JsonRpcProvider) => {
+    const provider = this.chainProviders[chainId];
+    if (!provider) {
+      return Result.fail(new ChainError(ChainError.reasons.ProviderNotFound));
+    }
+    return await this.retryWrapper<BigNumber>(chainId, async () => {
       let gasPrice: BigNumber | undefined = undefined;
       if (chainId === 1) {
         try {
@@ -443,16 +498,24 @@ export class EthereumChainReader implements IVectorChainReader {
       if (!gasPrice) {
         try {
           gasPrice = await provider.getGasPrice();
+          gasPrice = gasPrice.add(gasPrice.mul(BUMP_GAS_PRICE).div(100));
         } catch (e) {
           return Result.fail(e);
         }
+      }
+      if (gasPrice.lt(MIN_GAS_PRICE)) {
+        gasPrice = BigNumber.from(MIN_GAS_PRICE);
       }
       return Result.ok(gasPrice);
     });
   }
 
   async estimateGas(chainId: number, transaction: TransactionRequest): Promise<Result<BigNumber, ChainError>> {
-    return await this.retryWrapper<BigNumber>(chainId, async (provider: JsonRpcProvider) => {
+    const provider = this.chainProviders[chainId];
+    if (!provider) {
+      return Result.fail(new ChainError(ChainError.reasons.ProviderNotFound));
+    }
+    return await this.retryWrapper<BigNumber>(chainId, async () => {
       try {
         const gas = await provider.estimateGas(transaction);
         return Result.ok(gas);
@@ -468,7 +531,11 @@ export class EthereumChainReader implements IVectorChainReader {
     spender: string,
     chainId: number,
   ): Promise<Result<BigNumber, ChainError>> {
-    return await this.retryWrapper<BigNumber>(chainId, async (provider: JsonRpcProvider) => {
+    const provider = this.chainProviders[chainId];
+    if (!provider) {
+      return Result.fail(new ChainError(ChainError.reasons.ProviderNotFound));
+    }
+    return await this.retryWrapper<BigNumber>(chainId, async () => {
       const erc20 = new Contract(tokenAddress, ERC20Abi, provider);
       try {
         const res = await erc20.allowance(owner, spender);
@@ -480,7 +547,11 @@ export class EthereumChainReader implements IVectorChainReader {
   }
 
   async getOnchainBalance(assetId: string, balanceOf: string, chainId: number): Promise<Result<BigNumber, ChainError>> {
-    return await this.retryWrapper<BigNumber>(chainId, async (provider: JsonRpcProvider) => {
+    const provider = this.chainProviders[chainId];
+    if (!provider) {
+      return Result.fail(new ChainError(ChainError.reasons.ProviderNotFound));
+    }
+    return await this.retryWrapper<BigNumber>(chainId, async () => {
       try {
         const onchainBalance =
           assetId === AddressZero
@@ -494,7 +565,11 @@ export class EthereumChainReader implements IVectorChainReader {
   }
 
   async getDecimals(assetId: string, chainId: number): Promise<Result<number, ChainError>> {
-    return await this.retryWrapper<number>(chainId, async (provider: JsonRpcProvider) => {
+    const provider = this.chainProviders[chainId];
+    if (!provider) {
+      return Result.fail(new ChainError(ChainError.reasons.ProviderNotFound));
+    }
+    return await this.retryWrapper<number>(chainId, async () => {
       try {
         const decimals = assetId === AddressZero ? 18 : await new Contract(assetId, ERC20Abi, provider).decimals();
         return Result.ok(decimals);
@@ -509,7 +584,11 @@ export class EthereumChainReader implements IVectorChainReader {
     channelAddress: string,
     chainId: number,
   ): Promise<Result<boolean, ChainError>> {
-    return await this.retryWrapper<boolean>(chainId, async (provider: JsonRpcProvider) => {
+    const provider = this.chainProviders[chainId];
+    if (!provider) {
+      return Result.fail(new ChainError(ChainError.reasons.ProviderNotFound));
+    }
+    return await this.retryWrapper<boolean>(chainId, async () => {
       // check if it was deployed
       const code = await this.getCode(channelAddress, chainId);
       if (code.isError) {
@@ -554,7 +633,11 @@ export class EthereumChainReader implements IVectorChainReader {
   // }
 
   async registerChannel(channelAddress: string, chainId: number): Promise<Result<void, ChainError>> {
-    return this.retryWrapper<void>(chainId, async (provider: JsonRpcProvider) => {
+    const provider = this.chainProviders[chainId];
+    if (!provider) {
+      return Result.fail(new ChainError(ChainError.reasons.ProviderNotFound));
+    }
+    return this.retryWrapper<void>(chainId, async () => {
       if (this.contracts.has(channelAddress)) {
         // channel is already registered
         return Result.ok(undefined);
@@ -723,50 +806,48 @@ export class EthereumChainReader implements IVectorChainReader {
     chainId: number,
     bytecode?: string,
   ): Promise<Result<RegisteredTransfer[], ChainError>> {
-    return await this.retryWrapper<RegisteredTransfer[]>(chainId, async (provider: JsonRpcProvider) => {
-      // Registry for chain not loaded, load into memory
-      const registry = new Contract(transferRegistry, TransferRegistry.abi, provider);
-      let registered;
-      if (bytecode) {
-        // Try with evm first
-        const evm = this.tryEvm(registry.interface.encodeFunctionData("getTransferDefinitions"), bytecode);
-
-        if (!evm.isError) {
-          try {
-            registered = registry.interface.decodeFunctionResult("getTransferDefinitions", evm.getValue()!)[0];
-          } catch (e) {}
-        }
-      }
-      if (!registered) {
-        try {
-          registered = await registry.getTransferDefinitions();
-        } catch (e) {
-          return Result.fail(new ChainError(e.message, { chainId, transferRegistry }));
-        }
-      }
-      const cleaned = registered.map((r: RegisteredTransfer) => {
-        return {
-          name: r.name,
-          definition: r.definition,
-          stateEncoding: tidy(r.stateEncoding),
-          resolverEncoding: tidy(r.resolverEncoding),
-          encodedCancel: r.encodedCancel,
-        };
-      });
-      this.transferRegistries.set(chainId.toString(), cleaned);
-      return Result.ok(cleaned);
-    });
-  }
-
-  private async retryWrapper<T>(
-    chainId: number,
-    targetMethod: (provider: JsonRpcProvider) => Promise<Result<T, ChainError>>,
-  ): Promise<Result<T, ChainError>> {
     const provider = this.chainProviders[chainId];
     if (!provider) {
       return Result.fail(new ChainError(ChainError.reasons.ProviderNotFound));
     }
-    let res = await targetMethod(provider);
+    // Registry for chain not loaded, load into memory
+    const registry = new Contract(transferRegistry, TransferRegistry.abi, provider);
+    let registered;
+    if (bytecode) {
+      // Try with evm first
+      const evm = this.tryEvm(registry.interface.encodeFunctionData("getTransferDefinitions"), bytecode);
+
+      if (!evm.isError) {
+        try {
+          registered = registry.interface.decodeFunctionResult("getTransferDefinitions", evm.getValue()!)[0];
+        } catch (e) {}
+      }
+    }
+    if (!registered) {
+      try {
+        registered = await registry.getTransferDefinitions();
+      } catch (e) {
+        return Result.fail(new ChainError(e.message, { chainId, transferRegistry }));
+      }
+    }
+    const cleaned = registered.map((r: RegisteredTransfer) => {
+      return {
+        name: r.name,
+        definition: r.definition,
+        stateEncoding: tidy(r.stateEncoding),
+        resolverEncoding: tidy(r.resolverEncoding),
+        encodedCancel: r.encodedCancel,
+      };
+    });
+    this.transferRegistries.set(chainId.toString(), cleaned);
+    return Result.ok(cleaned);
+  }
+
+  private async retryWrapper<T>(
+    chainId: number,
+    targetMethod: () => Promise<Result<T, ChainError>>,
+  ): Promise<Result<T, ChainError>> {
+    let res = await targetMethod();
     let retries;
     const errors: { [attempt: number]: string | undefined } = {};
     if (!res.isError) {
@@ -775,7 +856,7 @@ export class EthereumChainReader implements IVectorChainReader {
 
     errors[0] = res.getError()?.message;
     for (retries = 1; retries < ETH_READER_MAX_RETRIES; retries++) {
-      res = await targetMethod(provider);
+      res = await targetMethod();
       if (!res.isError) {
         break;
       }